Rename AgendaLab2 to AgendaLab (forgot to stage my files)
