+++ PowerPointLabs/PowerPointLabs/Models/PowerPointAckSlide.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.IO;
