+++ PowerPointLabs/PowerPointLabs/ZoomToArea.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using PowerPointLabs.Models;
 using Office = Microsoft.Office.Core;
